src/oncology/requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# OpenAI/OpenRouter dependencies
openai>=1.3.7
anthropic>=0.5.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0

# Data processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# Medical data processing
pydicom>=2.4.0  # For DICOM medical images
nibabel>=5.1.0  # For neuroimaging files
vcfpy>=0.13.6   # For genomic VCF files

# NLP and text processing
transformers>=4.35.0
nltk>=3.8.1
spacy>=3.7.0

# Machine learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization (for debugging/analysis)
matplotlib>=3.8.0
seaborn>=0.13.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Monitoring and logging
tensorboard>=2.14.0
wandb>=0.15.0

# Web interface (if needed)
flask>=3.0.0
gunicorn>=21.2.0

# Performance optimization
numba>=0.58.0